Cleanup + Additions
Cleaned up a few code errors that we forgot about such as leaving an enable field on motors. Also added new features like dynamic memory for the simulator w/ pointers and added the simulator itself with some trig functions augmented into it for an accurate display of the car with a heading other than 0 deg.